+++ src/test/kotlin/dataSource/utils/FakeDataSource.kt
@@ -362,3 +362,224 @@ class FakeDataSource : CostOfLivingDataSource {
         )
     }
 }
+
+
+class FakeDataSourceM : CostOfLivingDataSource {
+    override fun getAllCitiesData(): List<CityEntity> {
+        return listOf(
+            CityEntity(
+                cityName = "Santiago de Cuba",
+                country = "Cuba",
+                mealsPrices = MealsPrices(
+                    mealInexpensiveRestaurant = 6.5f,
+                    mealFor2PeopleMidRangeRestaurant = 15.0f,
+                    mealAtMcDonaldSOrEquivalent = 5.0f
+                ),
+                drinksPrices = DrinksPrices(
+                    cappuccinoRegularInRestaurants = 2.0f,
+                    cokePepsiAThirdOfLiterBottleInRestaurants = 2.0f,
+                    waterAThirdOfLiterBottleInRestaurants = 1.0f,
+                    milkRegularOneLiter = 3.8f,
+                    waterOneAndHalfLiterBottleAtTheMarket = 2.0f
+                ),
+                fruitAndVegetablesPrices = FruitAndVegetablesPrices(
+                    apples1kg = 11.02f,
+                    banana1kg = 0.44f,
+                    oranges1kg = 1.1f,
+                    tomato1kg = 0.88f,
+                    potato1kg = 1.1f,
+                    onion1kg = 2.2f,
+                    lettuceOneHead = 0.25f
+                ),
+                foodPrices = FoodPrices(
+                    loafOfFreshWhiteBread500g = 1.1f,
+                    riceWhite1kg = 1.1f,
+                    eggsRegular12 = 2.03f,
+                    localCheese1kg = 9.0f,
+                    chickenFillets1kg = 6.61f,
+                    beefRound1kgOrEquivalentBackLegRedMeat = 6.61f
+                ),
+                servicesPrices = ServicesPrices(
+                    basicElectricityHeatingCoolingWaterGarbageFor85m2Apartment = null,
+                    oneMinOfPrepaidMobileTariffLocalNoDiscountsOrPlans = 0.1f,
+                    internet60MbpsOrMoreUnlimitedDataCableAdsl = null,
+                    fitnessClubMonthlyFeeForOneAdult = null,
+                    tennisCourtRentOneHourOnWeekend = null,
+                    cinemaInternationalReleaseOneSeat = 1.0f,
+                    preschoolOrKindergartenFullDayPrivateMonthlyForOneChild = null,
+                    internationalPrimarySchoolYearlyForOneChild = null
+                ),
+                clothesPrices = ClothesPrices(
+                    onePairOfJeansLevis50oneOrSimilar = 40.0f,
+                    oneSummerDressInAChainStoreZaraHAndM = 40.0f,
+                    onePairOfNikeRunningShoesMidRange = 60.0f,
+                    onePairOfMenLeatherBusinessShoes = 50.0f
+                ),
+                transportationsPrices = TransportationsPrices(
+                    oneWayTicketLocalTransport = null,
+                    monthlyPassRegularPrice = null,
+                    taxiStartNormalTariff = null,
+                    taxi1kmNormalTariff = 0.7f,
+                    taxi1hourWaitingNormalTariff = null,
+                    gasolineOneLiter = null
+                ),
+                carsPrices = CarsPrices(
+                    volkswagenGolf_1_4_90kwTrendLineOrEquivalentNewCar = null,
+                    toyotaCorollaSedan_1_6l_97kwComfortOrEquivalentNewCar = null
+                ),
+                realEstatesPrices = RealEstatesPrices(
+                    apartmentOneBedroomInCityCentre = null,
+                    apartmentOneBedroomOutsideOfCentre = null,
+                    apartment3BedroomsInCityCentre = 600.0f,
+                    apartment3BedroomsOutsideOfCentre = null,
+                    pricePerSquareMeterToBuyApartmentInCityCentre = null,
+                    pricePerSquareMeterToBuyApartmentOutsideOfCentre = null
+                ),
+                averageMonthlyNetSalaryAfterTax = 18f,
+                dataQuality = false
+            ),
+            CityEntity(
+                cityName = "Sancti Spiritus",
+                country = "Cuba",
+                mealsPrices = MealsPrices(
+                    mealInexpensiveRestaurant = 7.0f,
+                    mealFor2PeopleMidRangeRestaurant = 25.0f,
+                    mealAtMcDonaldSOrEquivalent = 5.0f
+                ),
+                drinksPrices = DrinksPrices(
+                    cappuccinoRegularInRestaurants = null,
+                    cokePepsiAThirdOfLiterBottleInRestaurants = 1.0f,
+                    waterAThirdOfLiterBottleInRestaurants = 0.75f,
+                    milkRegularOneLiter = null,
+                    waterOneAndHalfLiterBottleAtTheMarket = 1.0f
+                ),
+                fruitAndVegetablesPrices = FruitAndVegetablesPrices(
+                    apples1kg = 3.0f,
+                    banana1kg = 0.5f,
+                    oranges1kg = 0.5f,
+                    tomato1kg = 1.0f,
+                    potato1kg = 2.0f,
+                    onion1kg = 2.0f,
+                    lettuceOneHead = null
+                ),
+                foodPrices = FoodPrices(
+                    loafOfFreshWhiteBread500g = null,
+                    riceWhite1kg = 1.0f,
+                    eggsRegular12 = 2.4f,
+                    localCheese1kg = 4.0f,
+                    chickenFillets1kg = 5.0f,
+                    beefRound1kgOrEquivalentBackLegRedMeat = 10.0f
+                ),
+                servicesPrices = ServicesPrices(
+                    basicElectricityHeatingCoolingWaterGarbageFor85m2Apartment = 30.0f,
+                    oneMinOfPrepaidMobileTariffLocalNoDiscountsOrPlans = null,
+                    internet60MbpsOrMoreUnlimitedDataCableAdsl = null,
+                    fitnessClubMonthlyFeeForOneAdult = null,
+                    tennisCourtRentOneHourOnWeekend = null,
+                    cinemaInternationalReleaseOneSeat = null,
+                    preschoolOrKindergartenFullDayPrivateMonthlyForOneChild = null,
+                    internationalPrimarySchoolYearlyForOneChild = null
+                ),
+                clothesPrices = ClothesPrices(
+                    onePairOfJeansLevis50oneOrSimilar = 25.0f,
+                    oneSummerDressInAChainStoreZaraHAndM = 15.0f,
+                    onePairOfNikeRunningShoesMidRange = 40.0f,
+                    onePairOfMenLeatherBusinessShoes = 30.0f
+                ),
+                transportationsPrices = TransportationsPrices(
+                    oneWayTicketLocalTransport = null,
+                    monthlyPassRegularPrice = null,
+                    taxiStartNormalTariff = 1.0f,
+                    taxi1kmNormalTariff = 1.0f,
+                    taxi1hourWaitingNormalTariff = null,
+                    gasolineOneLiter = 2.0f
+                ),
+                carsPrices = CarsPrices(
+                    volkswagenGolf_1_4_90kwTrendLineOrEquivalentNewCar = 50000.0f,
+                    toyotaCorollaSedan_1_6l_97kwComfortOrEquivalentNewCar = 70000.0f
+                ),
+                realEstatesPrices = RealEstatesPrices(
+                    apartmentOneBedroomInCityCentre = null,
+                    apartmentOneBedroomOutsideOfCentre = null,
+                    apartment3BedroomsInCityCentre = null,
+                    apartment3BedroomsOutsideOfCentre = null,
+                    pricePerSquareMeterToBuyApartmentInCityCentre = null,
+                    pricePerSquareMeterToBuyApartmentOutsideOfCentre = null
+                ),
+                averageMonthlyNetSalaryAfterTax = 20.0f,
+                dataQuality = false
+            ),
+            CityEntity(
+                cityName = "Santa Clara",
+                country = "Cuba",
+                mealsPrices = MealsPrices(
+                    mealInexpensiveRestaurant = 10.0f,
+                    mealFor2PeopleMidRangeRestaurant = 30.0f,
+                    mealAtMcDonaldSOrEquivalent = 5.0f
+                ),
+                drinksPrices = DrinksPrices(
+                    cappuccinoRegularInRestaurants = 1.91f,
+                    cokePepsiAThirdOfLiterBottleInRestaurants = 2.0f,
+                    waterAThirdOfLiterBottleInRestaurants = 0.5f,
+                    milkRegularOneLiter = 3.52f,
+                    waterOneAndHalfLiterBottleAtTheMarket = 0.98f
+                ),
+                fruitAndVegetablesPrices = FruitAndVegetablesPrices(
+                    apples1kg = 6.24f,
+                    banana1kg = null,
+                    oranges1kg = 1.12f,
+                    tomato1kg = 1.0f,
+                    potato1kg = 1.82f,
+                    onion1kg = 2.0f,
+                    lettuceOneHead = 0.86f
+                ),
+                foodPrices = FoodPrices(
+                    loafOfFreshWhiteBread500g = 1.0f,
+                    riceWhite1kg = 1.0f,
+                    eggsRegular12 = 1.64f,
+                    localCheese1kg = 4.0f,
+                    chickenFillets1kg = 6.07f,
+                    beefRound1kgOrEquivalentBackLegRedMeat = 10.0f
+                ),
+                servicesPrices = ServicesPrices(
+                    basicElectricityHeatingCoolingWaterGarbageFor85m2Apartment = 11.5f,
+                    oneMinOfPrepaidMobileTariffLocalNoDiscountsOrPlans = 0.26f,
+                    internet60MbpsOrMoreUnlimitedDataCableAdsl = 80.0f,
+                    fitnessClubMonthlyFeeForOneAdult = 6.0f,
+                    tennisCourtRentOneHourOnWeekend = null,
+                    cinemaInternationalReleaseOneSeat = 1.12f,
+                    preschoolOrKindergartenFullDayPrivateMonthlyForOneChild = 70.0f,
+                    internationalPrimarySchoolYearlyForOneChild = 15166.67f
+                ),
+                clothesPrices = ClothesPrices(
+                    onePairOfJeansLevis50oneOrSimilar = 55.5f,
+                    oneSummerDressInAChainStoreZaraHAndM = 44.33f,
+                    onePairOfNikeRunningShoesMidRange = 79.17f,
+                    onePairOfMenLeatherBusinessShoes = 78.0f
+                ),
+                transportationsPrices = TransportationsPrices(
+                    oneWayTicketLocalTransport = 0.5f,
+                    monthlyPassRegularPrice = null,
+                    taxiStartNormalTariff = 1.0f,
+                    taxi1kmNormalTariff = 0.5f,
+                    taxi1hourWaitingNormalTariff = 15.0f,
+                    gasolineOneLiter = 1.14f
+                ),
+                carsPrices = CarsPrices(
+                    volkswagenGolf_1_4_90kwTrendLineOrEquivalentNewCar = 80000.0f,
+                    toyotaCorollaSedan_1_6l_97kwComfortOrEquivalentNewCar = 81500.0f
+                ),
+                realEstatesPrices = RealEstatesPrices(
+                    apartmentOneBedroomInCityCentre = null,
+                    apartmentOneBedroomOutsideOfCentre = null,
+                    apartment3BedroomsInCityCentre = null,
+                    apartment3BedroomsOutsideOfCentre = null,
+                    pricePerSquareMeterToBuyApartmentInCityCentre = null,
+                    pricePerSquareMeterToBuyApartmentOutsideOfCentre = null
+                ),
+                averageMonthlyNetSalaryAfterTax = 25.0f,
+                dataQuality = false
+            )
+        )
+    }
+}
